requirements.txt
pandas~=2.0.0
fitfile~=0.0.1
pip~=20.0.2
colorama~=0.4.6
platformdirs~=3.2.0
tox~=4.4.11
tomli~=2.0.1
pluggy~=1.0.0
packaging~=23.0
chardet~=5.1.0
virtualenv~=20.21.0
cachetools~=5.3.0
filelock~=3.10.7
pytest~=7.2.2
parso~=0.8.3
jedi~=0.18.2
mypy~=1.1.1
pytz~=2023.3
attrs~=22.2.0
setuptools~=44.0.0
numpy~=1.24.2
wheel~=0.34.2
openpyxl~=3.1.2
python-dateutil~=2.8.2
Pygments~=2.14.0
ipython~=8.12.0
traitlets~=5.9.0
decorator~=5.1.1
backcall~=0.2.0
pickleshare~=0.7.5
pexpect~=4.8.0
wcwidth~=0.2.6
exceptiongroup~=1.1.1
iniconfig~=2.0.0
ptyprocess~=0.7.0
six~=1.16.0
asttokens~=2.2.1
executing~=1.2.0
distlib~=0.3.6